Separated the navigation tree setup to its own location
